pyproject.toml
[tool.poetry]
name = "evalyn-bot"
version = "1.0.0"
description = "Telegram Bot Evalyn Era"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"